added dateHelpers file to share date converstion
